pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/storemate_report_generator"]

[project]
name = "storemate_report_generator"
version = "0.1.0"
authors = [
    { name = "kasrsf", email = "kasra@kasrsf.com" },
]
description = "A tool for processing store DBF data and generating analytical reports"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dbfread>=2.0.7",
    "duckdb>=1.0.0",
    "PyYAML>=6.0",
    "click>=8.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.10"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=src"

[tool.ruff]
line-length = 100
target-version = "py39"
lint.select = ["E", "F", "B", "I", "N", "UP", "S", "C",]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "B017"]
